+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * IMPORTANT Fix captcha form issue [#132](https://github.com/serphacker/serposcope/issues/132)
 * Fix NPE in log anonymizer [#127](https://github.com/serphacker/serposcope/issues/127)
-* 
+* Fix mobile icon no more displayed [#128](https://github.com/serphacker/serposcope/issues/128)
 
 ## 2.5.1 - 2016-07-10